Dataset Skara Scen 1_ver3
Did an update because the Lidköping-model induced new changes to the underlying vensim structure, however, do not think it had any effect on the Skara dataset. Anyways, I wanted to check how the order of the columns came out. It seems at Default the Skara and Lidköping outputs come out in different order when it comes to the columns. Hence, a good plan to create directories in Python that are independent of the order of the column as you are planning.