Complete Sankey data set Skara and Lidköping
There is an issue worth noting, and it is the number of values each data point has. If you in your code reduce the amount of decimals this is perfectly fine.
Also, for this reason, there is an issue with scenario 4 for both Skara and Lidköping, where the input slider value is set to "0.96" but results in "0.959999996" for some unclear reason. This issue needs to be bridged in the Python code somehow. //Gary